Silence warnings.

+++ ext/pdo_pgsql/pgsql_driver.c
@@ -16,7 +16,7 @@
   +----------------------------------------------------------------------+
 */
 
-/* $Id: pgsql_driver.c,v 1.42 2005/06/28 14:54:44 iliaa Exp $ */
+/* $Id: pgsql_driver.c,v 1.43 2005/06/30 00:07:42 iliaa Exp $ */
 
 #ifdef HAVE_CONFIG_H
 #include "config.h"
@@ -214,7 +214,7 @@ static char *pdo_pgsql_last_insert_id(pdo_dbh_t *dbh, const char *name, unsigned
 
 	/* TODO: if name != NULL, pull out last value for that sequence/column */
 
-	*len = spprintf(&id, 0, "%ld", H->pgoid);
+	*len = spprintf(&id, 0, "%ld", (long) H->pgoid);
 	return id;
 }
